src/utils/requirements.txt
# Bibliotecas de data science
numpy>=1.19.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Deep Learning
tensorflow>=2.8.0
tensorflow-addons>=0.16.0

# Processamento de imagem
pillow>=9.0.0
opencv-python>=4.5.5
scipy>=1.8.0

# MLOps
mlflow>=1.20.0

# Utilitários
jupyter>=1.0.0
python-dotenv>=0.19.0
tqdm>=4.62.0
ipykernel>=6.0.0
pyyaml>=6.0

# Visualização avançada
plotly>=5.5.0
